requirements.txt
ultralytics
numpy
matplotlib
polars
pyyaml
Pillow
psutil
requests>=2.23.0
torch>=1.8.0 
torchvision>=0.9.0 
opencv-python
pyodbc
pyautogui
